Factor out common functionality
Extract common functionality present in both implementations for interacting with the shared lib
