+++ internal/shared_lib_core_unix.go
@@ -3,14 +3,9 @@
 package internal
 
 import (
-	"context"
 	"encoding/json"
 	"errors"
 	"fmt"
-	"log"
-	"os"
-	"path"
-	"runtime"
 	"unsafe"
 )
 
@@ -85,60 +80,6 @@ type SharedLibCore struct {
 
 var coreLib *SharedLibCore
 
-// find1PasswordLibPath returns the path to the 1Password shared library
-// (libop_sdk_ipc_client.dylib/.so/.dll) depending on OS.
-func find1PasswordLibPath() (string, error) {
-	var locations []string
-
-	home, err := os.UserHomeDir()
-	if err != nil {
-		return "", err
-	}
-
-	switch runtime.GOOS {
-	case "darwin":
-		locations = []string{
-			"/Applications/1Password.app/Contents/Frameworks/libop_sdk_ipc_client.dylib",
-			path.Join(home, "Applications/1Password.app/Contents/Frameworks/libop_sdk_ipc_client.dylib"),
-		}
-
-	case "linux":
-		locations = []string{
-			"/usr/bin/1password/libop_sdk_ipc_client.so",
-			"/opt/1Password/libop_sdk_ipc_client.so",
-			"/snap/bin/1password/libop_sdk_ipc_client.so",
-		}
-
-	default:
-		return "", fmt.Errorf("unsupported OS: %s", runtime.GOOS)
-	}
-	for _, libPath := range locations {
-		if _, err := os.Stat(libPath); err == nil {
-			return libPath, nil
-		}
-	}
-
-	return "", fmt.Errorf("1Password desktop application not found")
-}
-
-func GetSharedLibCore(accountName string) (*CoreWrapper, error) {
-	if coreLib == nil {
-		path, err := find1PasswordLibPath()
-		if err != nil {
-			return nil, err
-		}
-		coreLib, err = loadCore(path)
-		if err != nil {
-			return nil, err
-		}
-		coreLib.accountName = accountName
-	}
-
-	coreWrapper := CoreWrapper{InnerCore: coreLib}
-
-	return &coreWrapper, nil
-}
-
 func loadCore(path string) (*SharedLibCore, error) {
 	cPath := C.CString(path)
 	defer C.free(unsafe.Pointer(cPath))
@@ -171,56 +112,6 @@ func loadCore(path string) (*SharedLibCore, error) {
 		sendMessage:  (C.send_message_t)(fnSend),
 		freeResponse: (C.free_message_t)(fnFree),
 	}, nil
-}
-
-// InitClient creates a client instance in the current core module and returns its unique ID.
-func (slc *SharedLibCore) InitClient(ctx context.Context, config []byte) ([]byte, error) {
-	const kind = "init_client"
-	request := Request{
-		Kind:        kind,
-		AccountName: slc.accountName,
-		Payload:     config,
-	}
-
-	requestMarshaled, err := json.Marshal(request)
-	if err != nil {
-		return nil, err
-	}
-	res, err := slc.callSharedLibrary(requestMarshaled)
-	if err != nil {
-		return nil, err
-	}
-
-	return res, nil
-}
-
-func (slc *SharedLibCore) Invoke(ctx context.Context, invokeConfig []byte) ([]byte, error) {
-	const kind = "invoke"
-	request := Request{
-		Kind:        kind,
-		AccountName: slc.accountName,
-		Payload:     invokeConfig,
-	}
-
-	requestMarshaled, err := json.Marshal(request)
-	if err != nil {
-		return nil, err
-	}
-
-	res, err := slc.callSharedLibrary(requestMarshaled)
-	if err != nil {
-		return nil, err
-	}
-
-	return res, nil
-}
-
-// ReleaseClient releases memory in the core associated with the given client ID.
-func (slc *SharedLibCore) ReleaseClient(clientID []byte) {
-	_, err := slc.callSharedLibrary(clientID)
-	if err != nil {
-		log.Println("failed to release client")
-	}
 }
 
 func (slc *SharedLibCore) callSharedLibrary(input []byte) ([]byte, error) {
@@ -260,19 +151,4 @@ func (slc *SharedLibCore) callSharedLibrary(input []byte) ([]byte, error) {
 	} else {
 		return nil, response
 	}
-}
-
-type Request struct {
-	Kind        string `json:"kind"`
-	AccountName string `json:"account_name"`
-	Payload     []byte `json:"payload"`
-}
-
-type Response struct {
-	Success bool   `json:"success"`
-	Payload []byte `json:"payload"`
-}
-
-func (r Response) Error() string {
-	return string(r.Payload)
 }
